Token fetching classes moved to an appropriate package
