+++ token-repository/src/test/scala/ch/datascience/tokenrepository/repository/fetching/interpreters.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2019 Swiss Data Science Center (SDSC)
+ * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+ * Eidgenössische Technische Hochschule Zürich (ETHZ).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ch.datascience.tokenrepository.repository.fetching
+
+import cats.implicits._
+import cats.effect.IO
+import ch.datascience.db.TransactorProvider
+import ch.datascience.tokenrepository.repository.AccessTokenCrypto
+
+import scala.util.Try
+
+private class TryPersistedTokensFinder(transactorProvider: TransactorProvider[Try])
+    extends PersistedTokensFinder[Try](transactorProvider)
+
+private class IOTokenFinder(
+    tokenInRepoFinder: PersistedTokensFinder[IO],
+    accessTokenCrypto: AccessTokenCrypto[IO]
+) extends TokenFinder[IO](tokenInRepoFinder, accessTokenCrypto)
